feat: Add module bootstrap script and update structure
This commit adds a new module bootstrap script and updates the structure of the repository. The module bootstrap script installs the 'Sampler' module if it's not already installed and then uses 'Invoke-Plaster' to generate files based on a template. The structure update includes moving files and directories using the 'Move-Files' function. These changes are necessary to improve the organization and setup of the repository.

Co-authored-by: Anthony Watherston <Anthony.Watherston@microsoft.com>